requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in' '.\requirements_test.in'
#
absl-py==2.1.0
    # via rouge-score
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via httpx
bert-score==0.3.13
    # via -r .\requirements_test.in
blis==0.7.11
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   nltk
    #   typer
cloudpathlib==0.20.0
    # via weasel
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   wasabi
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.1
    # via matplotlib
cupy-cuda12x==12.3.0
    # via spacy
cycler==0.12.1
    # via matplotlib
cymem==2.0.8
    # via
    #   preshed
    #   spacy
    #   thinc
faiss-cpu==1.9.0
    # via -r .\requirements.in
fastrlock==0.8.2
    # via cupy-cuda12x
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.55.1
    # via matplotlib
fsspec==2024.9.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via ollama
huggingface-hub==0.26.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   spacy
    #   torch
joblib==1.4.2
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.4.7
    # via matplotlib
langcodes==3.4.1
    # via spacy
language-data==1.2.0
    # via langcodes
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.3
    # via bert-score
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
murmurhash==1.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
neo4j==5.26.0
    # via
    #   -r .\requirements.in
    #   neo4j-graphrag
neo4j-graphrag==1.2.1
    # via -r .\requirements.in
networkx==3.4.2
    # via torch
nltk==3.9.1
    # via
    #   -r .\requirements_test.in
    #   rouge-score
numpy==1.26.4
    # via
    #   -r .\requirements.in
    #   bert-score
    #   blis
    #   contourpy
    #   cupy-cuda12x
    #   faiss-cpu
    #   matplotlib
    #   pandas
    #   rouge-score
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
    #   transformers
ollama==0.4.1
    # via -r .\requirements.in
packaging==24.1
    # via
    #   bert-score
    #   faiss-cpu
    #   huggingface-hub
    #   matplotlib
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pandas==2.2.3
    # via
    #   -r .\requirements.in
    #   bert-score
pillow==11.0.0
    # via
    #   matplotlib
    #   sentence-transformers
preshed==3.0.9
    # via
    #   spacy
    #   thinc
pyarrow==18.0.0
    # via -r .\requirements.in
pydantic==2.9.2
    # via
    #   confection
    #   neo4j-graphrag
    #   ollama
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via
    #   neo4j
    #   pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.9.11
    # via
    #   nltk
    #   transformers
requests==2.32.3
    # via
    #   bert-score
    #   huggingface-hub
    #   spacy
    #   transformers
    #   weasel
rich==13.9.4
    # via typer
rouge-score==0.1.2
    # via -r .\requirements_test.in
safetensors==0.4.5
    # via transformers
scikit-learn==1.5.2
    # via
    #   -r .\requirements_test.in
    #   sentence-transformers
scipy==1.14.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.2.1
    # via -r .\requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   python-dateutil
    #   rouge-score
smart-open==7.0.5
    # via weasel
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
spacy[cuda12x]==3.7.5
    # via -r .\requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.4.8
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
sympy==1.13.1
    # via torch
thinc==8.2.5
    # via spacy
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.20.3
    # via transformers
torch==2.5.1
    # via
    #   -r .\requirements_test.in
    #   bert-score
    #   sentence-transformers
tqdm==4.66.6
    # via
    #   bert-score
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   spacy
    #   transformers
transformers==4.46.2
    # via
    #   -r .\requirements_test.in
    #   bert-score
    #   sentence-transformers
typer==0.12.5
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.16.0
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
